+++ docs/client/components/pages/Undo/CustomUndoEditor/index.js
@@ -5,17 +5,20 @@ import Editor from 'draft-js-plugins-editor';
 import createUndoPlugin from 'draft-js-undo-plugin';
 import editorStyles from './editorStyles.css';
 import buttonStyles from './buttonStyles.css';
+import historyStyles from './historyStyles.css';
 
 const theme = Map({
   undo: buttonStyles.button,
   redo: buttonStyles.button,
+  historyItem: historyStyles.historyItem,
+  historyItemActive: historyStyles.historyItemActive,
 });
 const undoPlugin = createUndoPlugin({
   undoContent: 'Undo',
   redoContent: 'Redo',
   theme,
 });
-const { UndoButton, RedoButton } = undoPlugin;
+const { UndoButton, RedoButton, History } = undoPlugin;
 
 export default class CustomUndoEditor extends Component {
 
@@ -52,6 +55,7 @@ export default class CustomUndoEditor extends Component {
             editorState={ this.state.editorState }
             onChange={ this.onChange }
           />
+          <History editorState={ this.state.editorState } />
         </div>
       </div>
     );