+++ draft-js-undo-plugin/src/index.js
@@ -1,5 +1,6 @@
 import UndoButton from './UndoButton';
 import RedoButton from './RedoButton';
+import History from './History';
 import { Map } from 'immutable';
 import styles from './styles.css';
 import decorateComponentWithProps from 'decorate-component-with-props';
@@ -7,6 +8,8 @@ import decorateComponentWithProps from 'decorate-component-with-props';
 const defaultTheme = Map({
   redo: styles.button,
   undo: styles.button,
+  historyItem: styles.historyItem,
+  historyItemActive: styles.historyItemActive,
 });
 
 const historyPlugin = (config = {}) => {
@@ -23,6 +26,7 @@ const historyPlugin = (config = {}) => {
   return {
     UndoButton: decorateComponentWithProps(UndoButton, { theme, children: undoContent }),
     RedoButton: decorateComponentWithProps(RedoButton, { theme, children: redoContent }),
+    History: decorateComponentWithProps(History, { theme }),
   };
 };