+++ draft-js-undo-plugin/CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## To Be Released
 
+### Added
+
+- Exposed the History component
+
 ## 0.0.5 - 2016-03-25
 ### Released the first working of DraftJS Undo Plugin